# Conflicts:
#	resources/views/admin/index.blade.php
#	routes/web.php

+++ app/Jobs/SyncRoster.php
@@ -7,31 +7,32 @@ use App\DTOs\VatusaFacilityInfoDTO;
 use App\DTOs\VatusaRosterUser;
 use App\Models\Staff;
 use App\Models\User;
+use Http;
 use Illuminate\Contracts\Broadcasting\ShouldBeUnique;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Queue\Queueable;
-use Http;
 use Illuminate\Support\Facades\Log;
 
-class SyncRoster implements ShouldQueue, ShouldBeUnique
+class SyncRoster implements ShouldBeUnique, ShouldQueue
 {
     use Queueable;
-    private readonly string $ROSTER_API_ENDPOINT;
-    private readonly string $FACILITY_INFO_ENDPOINT;
 
     /**
      * Create a new job instance.
      */
-    public function __construct()
-    {
-        $this->ROSTER_API_ENDPOINT = config('app.vatusa_api_url') . '/v2/facility/' . config('app.vatusa_facility') . '/roster/both';
-        $this->FACILITY_INFO_ENDPOINT = config('app.vatusa_api_url') . '/v2/facility/' . config('app.vatusa_facility');
-    }
+    public function __construct() {}
 
-    private function updateRoster() {
-        $rosterData = Http::get($this->ROSTER_API_ENDPOINT, [
-            'apikey' => config('app.vatusa_api_key')
+    private function updateRoster()
+    {
+        $ROSTER_API_ENDPOINT = config('app.vatusa_api_url').'/v2/facility/'.config('app.vatusa_facility').'/roster/both';
+
+        $rosterData = Http::get($ROSTER_API_ENDPOINT, [
+            'apikey' => config('app.vatusa_api_key'),
         ]);
+
+        if ($rosterData->failed()) {
+            throw new \Exception('Failed to fetch roster data: '.$rosterData->status().' - '.$rosterData->body());
+        }
 
         $roster = $rosterData->json();
         User::where(['rostered' => true])->update(['rostered' => false]);
@@ -44,13 +45,27 @@ class SyncRoster implements ShouldQueue, ShouldBeUnique
 
         // Clear hanging OIs
         User::where([
-            'rostered' => false
+            'rostered' => false,
         ])->update([
-            'operating_initials' => null
+            'operating_initials' => null,
         ]);
     }
 
-    private function clearUserRoles() {
+    private function syncRosteredRole()
+    {
+        $users = User::all();
+
+        foreach ($users as $user) {
+            if ($user->rostered) {
+                $user->assignRole('rostered');
+            } else {
+                $user->removeRole('rostered');
+            }
+        }
+    }
+
+    private function clearUserRoles()
+    {
         $users = User::all();
 
         foreach ($users as $user) {
@@ -58,7 +73,8 @@ class SyncRoster implements ShouldQueue, ShouldBeUnique
         }
     }
 
-    private function assignRoles() {
+    private function assignRoles()
+    {
         $staffMembers = Staff::all();
 
         foreach ($staffMembers as $staff) {
@@ -70,10 +86,11 @@ class SyncRoster implements ShouldQueue, ShouldBeUnique
                     $user?->assignRole('admin', 'training', 'instructor', 'facilities', 'events', 'staff');
                     break;
                 case 'TA':
+                case 'ATA':
                     $user?->assignRole('admin', 'training', 'staff');
                     break;
                 case 'WM':
-                    $user?->assignRole('staff', 'admin');
+                    $user?->assignRole('admin', 'training', 'instructor', 'facilities', 'events', 'staff');
                     break;
                 case 'EC':
                     $user?->assignRole('events', 'staff');
@@ -91,10 +108,16 @@ class SyncRoster implements ShouldQueue, ShouldBeUnique
         }
     }
 
-    private function updateStaffMembers() {
-        $facilityInfo = Http::get($this->FACILITY_INFO_ENDPOINT, [
-            'apikey' => config('app.vatusa_api_key')
+    private function updateStaffMembers()
+    {
+        $FACILITY_INFO_ENDPOINT = config('app.vatusa_api_url').'/v2/facility/'.config('app.vatusa_facility');
+        $facilityInfo = Http::get($FACILITY_INFO_ENDPOINT, [
+            'apikey' => config('app.vatusa_api_key'),
         ]);
+
+        if ($facilityInfo->failed()) {
+            throw new \Exception('Failed to fetch facility info: '.$facilityInfo->status().' - '.$facilityInfo->body());
+        }
 
         $this->clearUserRoles();
         Staff::truncate();
@@ -118,7 +141,7 @@ class SyncRoster implements ShouldQueue, ShouldBeUnique
 
             if (App::environment() == 'development') {
                 $testUsers = User::where([
-                    'first_name' => "Web"
+                    'first_name' => 'Web',
                 ])->get();
 
                 foreach ($testUsers as $user) {
@@ -130,10 +153,16 @@ class SyncRoster implements ShouldQueue, ShouldBeUnique
                 }
             }
 
+            $this->syncRosteredRole();
+
             Log::info('Roster sync completed successfully.');
         } catch (\Exception $e) {
             // Log error
-            Log::error('Error syncing roster: '.$e->getMessage());
+            Log::error('Error syncing roster: '.$e->getMessage().'\n'.$e->getTraceAsString(), [
+                'url' => config('app.vatusa_api_url').'/v2/facility/'.config('app.vatusa_facility'),
+                'environment' => App::environment(),
+                'exception' => get_class($e),
+            ]);
         }
     }
 }
